+++ lib/slaw/grammars/za/act_nodes.rb
@@ -1,4 +1,5 @@
 require 'slaw/grammars/core_nodes'
+require 'slaw/grammars/inlines_nodes'
 
 module Slaw
   module Grammars
@@ -83,7 +84,7 @@ module Slaw
               b.preface { |b|
                 statements.elements.each { |element|
                   for e in element.elements
-                    e.to_xml(b, "") if e.is_a? NakedStatement
+                    e.to_xml(b, "") if e.is_a? Slaw::Grammars::Inlines::NakedStatement
                   end
                 }
               }
@@ -268,58 +269,6 @@ module Slaw
                 end
               }
             }
-          end
-        end
-
-        class NakedStatement < Treetop::Runtime::SyntaxNode
-          def to_xml(b, idprefix, i=0)
-            b.p { |b| clauses.to_xml(b, idprefix) } if clauses
-          end
-
-          def content
-            clauses
-          end
-        end
-
-        class Clauses < Treetop::Runtime::SyntaxNode
-          def to_xml(b, idprefix=nil)
-            for e in elements
-              if e.respond_to? :to_xml
-                e.to_xml(b, idprefix)
-              else
-                b << e.text_value
-              end
-            end
-          end
-        end
-
-        class Remark < Treetop::Runtime::SyntaxNode
-          def to_xml(b, idprefix)
-            b.remark(status: 'editorial') do |b|
-              b << '['
-              for e in content.elements
-                if e.respond_to? :to_xml
-                  e.to_xml(b, idprefix)
-                else
-                  b << e.text_value
-                end
-              end
-              b << ']'
-            end
-          end
-        end
-
-        class Image < Treetop::Runtime::SyntaxNode
-          def to_xml(b, idprefix)
-            attrs = {src: href.text_value}
-            attrs[:alt] = content.text_value unless content.text_value.empty?
-            b.img(attrs)
-          end
-        end
-
-        class Ref < Treetop::Runtime::SyntaxNode
-          def to_xml(b, idprefix)
-            b.ref(content.text_value, href: href.text_value)
           end
         end
